agent_15_blog_geo/requirements.txt
# Blog GEO Analyzer Requirements

# This file contains the dependencies required for the Blog GEO Analyzer project.
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# database
pymysql>=1.1.0
SQLAlchemy>=2.0.0  

# web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# image
Pillow>=10.0.0
requests>=2.31.0

# utilities
python-dateutil>=2.8.2
typing-extensions>=4.8.0

# logging
coloredlogs>=15.0